# Conflicts:
#	pom.xml
#	src/main/java/ru/skypro/homework/controller/AdsController.java
#	src/main/java/ru/skypro/homework/controller/AuthController.java
#	src/main/java/ru/skypro/homework/controller/UserController.java
#	src/main/java/ru/skypro/homework/dto/AdsDto.java
#	src/main/java/ru/skypro/homework/dto/UserDto.java
#	src/main/java/ru/skypro/homework/entity/Ads.java
#	src/main/java/ru/skypro/homework/entity/AdsComment.java
#	src/main/java/ru/skypro/homework/entity/User.java
#	src/main/java/ru/skypro/homework/repository/AdsCommentRepository.java
#	src/main/java/ru/skypro/homework/repository/AdsRepository.java
#	src/main/java/ru/skypro/homework/repository/UserRepository.java
#	src/main/java/ru/skypro/homework/service/AdsService.java
#	src/main/java/ru/skypro/homework/service/AuthService.java
#	src/main/java/ru/skypro/homework/service/UserService.java
#	src/main/java/ru/skypro/homework/service/impl/AdsServiceImpl.java
#	src/main/java/ru/skypro/homework/service/impl/AuthServiceImpl.java
#	src/main/java/ru/skypro/homework/service/impl/UserServiceImpl.java
#	src/main/resources/application.properties

+++ src/main/java/ru/skypro/homework/repository/AdsCommentRepository.java
@@ -3,6 +3,6 @@ package ru.skypro.homework.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import ru.skypro.homework.entity.AdsComment;
 
-public interface AdsCommentRepository extends JpaRepository <AdsComment, Integer> {
+public interface AdsCommentRepository extends JpaRepository<AdsComment, Long> {
 
 }
